Trains a model given training data input_fn.

Trains a model given training data input_fn.

# S3 method for tf_custom_model
fit(object, input_fn, steps = NULL, hooks = NULL,
  max_steps = NULL)

Arguments

input_fn
Input function returning a list of: features - Tensor or dictionary of string feature name to Tensor. labels - Tensor or dictionary of Tensor with labels.

steps
Number of steps for which to train model. If NULL, train forever or train until input_fn generates the OutOfRange or StopIteration error. 'steps' works incrementally. If you call two times train(steps=10) then training occurs in total 20 steps. If OutOfRange or StopIteration error occurs in the middle, training stops before 20 steps. If you don't want to have incremental behaviour please set max_steps instead. If set, max_steps must be NULL.

hooks
List of SessionRunHook subclass instances. Used for callbacks inside the training loop.

max_steps
Number of total steps for which to train model. If NULL, train forever or train until input_fn generates the OutOfRange or StopIteration error. If set, steps must be NULL. If OutOfRange or StopIteration error occurs in the middle, training stops before max_steps steps.

Value

self, for chaining.

Raises

ValueError: If both steps and max_steps are not NULL. ValueError: If either steps or max_steps is <= 0.

See also

Other custom estimator: estimator_spec, estimator, evaluate.tf_custom_model, export_savedmodel.tf_custom_model, predict.tf_custom_model